# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WebForge AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  pro: 'pro',
  business: 'business',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  plan: UserPlan;
  projectsThisMonth: number;
  /** @nullable */
  projectLimit?: number | null;
  createdAt: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type ProjectType = typeof ProjectType[keyof typeof ProjectType];


export const ProjectType = {
  landing_page: 'landing_page',
  ecommerce: 'ecommerce',
  blog: 'blog',
  portfolio: 'portfolio',
  dashboard: 'dashboard',
  saas: 'saas',
  web_app: 'web_app',
} as const;

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  generating: 'generating',
  ready: 'ready',
  published: 'published',
  error: 'error',
} as const;

export interface Project {
  id: number;
  userId: number;
  name: string;
  description?: string;
  type: ProjectType;
  status: ProjectStatus;
  prompt: string;
  /** @nullable */
  generatedCode?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ProjectInputType = typeof ProjectInputType[keyof typeof ProjectInputType];


export const ProjectInputType = {
  landing_page: 'landing_page',
  ecommerce: 'ecommerce',
  blog: 'blog',
  portfolio: 'portfolio',
  dashboard: 'dashboard',
  saas: 'saas',
  web_app: 'web_app',
} as const;

export interface ProjectInput {
  name: string;
  description?: string;
  type: ProjectInputType;
  prompt: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  draft: 'draft',
  generating: 'generating',
  ready: 'ready',
  published: 'published',
  error: 'error',
} as const;

export interface ProjectUpdate {
  name?: string;
  description?: string;
  prompt?: string;
  status?: ProjectUpdateStatus;
}

export interface GenerateInput {
  prompt: string;
  regenerate?: boolean;
}

export type QuickGenerateInputType = typeof QuickGenerateInputType[keyof typeof QuickGenerateInputType];


export const QuickGenerateInputType = {
  landing_page: 'landing_page',
  ecommerce: 'ecommerce',
  blog: 'blog',
  portfolio: 'portfolio',
  dashboard: 'dashboard',
  saas: 'saas',
  web_app: 'web_app',
} as const;

export interface QuickGenerateInput {
  prompt: string;
  type?: QuickGenerateInputType;
}

export interface ProjectStructure {
  pages: string[];
  features: string[];
  techStack: string[];
  /** @nullable */
  dbSchema?: string | null;
  apiEndpoints: string[];
}

export interface GenerationResult {
  projectId: number;
  status: string;
  /** @nullable */
  generatedCode?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  structure?: ProjectStructure;
  message: string;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface DashboardStats {
  totalProjects: number;
  publishedProjects: number;
  generationsThisMonth: number;
  /** @nullable */
  planLimit?: number | null;
  planName: string;
  projectsByType: TypeCount[];
  recentGenerations: number;
}

export interface ActivityItem {
  id: number;
  projectId: number;
  projectName: string;
  action: string;
  timestamp: string;
  type: string;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  currency: string;
  interval: string;
  /** @nullable */
  projectLimit?: number | null;
  features: string[];
  highlighted: boolean;
}